+++ change_log.md
@@ -7,6 +7,37 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.7.0] - 2025-02-20
+
+### Added
+
+* There are now a bunch more themes available (`sakura`, `tacit`, `XP`, and more).
+* The deployment URLs for Skulpt and its associated files can now be overridden
+* 
+### Fixed
+
+* Deployed sites were not correctly escaping HTML entities for generated tests
+* Fixed an issue where the configuration parameters were not being accepted via start_server
+* Slightly better error message when the Bakery library is not found, and assert_equal is used
+* Theme architecture is more flexible and dynamically based off `libs/manifest.json`
+
+
+## [1.6.2] - 2024-12-09
+
+### Fixed
+
+* The options for a `SelectBox` are forced into strings, before they are escaped.
+
+## [1.6.1] - 2024-12-08
+
+### Fixed
+
+* Using utcnow() instead of now() (credit to @crsommers)
+* Refactored `Div` and `Span` slightly (credit to @crsommers)
+* Fixed issue with escaping of tags in `VisitedPage`
+* Fixed issue with dropdowns not properly escaping the option
+* Fixed issue with non-button related values being passed incorrectly
+
 ## [1.6.0] - 2024-12-02
 
 ### Added